Move eqK and genK test instances to arrow-test
