+++ modules/mtl/arrow-mtl-test/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Arrow MTL Test
+POM_ARTIFACT_ID=arrow-mtl-test
+POM_PACKAGING=jar
