+++ modules/mtl/arrow-mtl-test/src/main/kotlin/arrow/mtl/WriterT.kt
@@ -1,0 +1,9 @@
+package arrow.mtl
+
+import arrow.Kind
+import arrow.mtl.WriterT
+import arrow.mtl.WriterTPartialOf
+import arrow.test.generators.GenK
+import arrow.test.generators.tuple2
+import io.kotlintest.properties.Gen
+
